task_planner/requirements.txt
Django==5.2
requests==2.32.3
djangorestframework